requirements.txt
pymongo==4.8.0
qdrant-client==1.12.1
pdfplumber==0.11.4
sentence-transformers==3.2.1
groq==0.13.0
tqdm==4.66.5
langdetect==1.0.9
python-dotenv==1.0.1
pydantic==2.9.2
colorama==0.4.6
